+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to
 - Move generic oauth login provider to `pkg/kiosk/login/goauth` package with public `Run()` entrypoint
 - Move idtoken login provider to `pkg/kiosk/login/idtoken` package with public `Run()` entrypoint
 - Move apikey login provider to `pkg/kiosk/login/apikey` package with public `Run()` entrypoint
+- Move aws login provider to `pkg/kiosk/login/aws` package with public `Run()` entrypoint
 
 ### CI/CD